Fix jbdc

+++ DATN/src/main/java/com/huutuan/bitmap/Controller/UserController.java
@@ -40,7 +40,7 @@ public class UserController {
 	}
 
 	@GetMapping("/edit/{id}")
-	public String showUpdateForm(@PathVariable("id") long id, Model model) {
+	public String showUpdateForm(@PathVariable("id") int id, Model model) {
 		User user = userRepository.findById(id)
 				.orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
 		model.addAttribute("user", user);
@@ -60,7 +60,7 @@ public class UserController {
 	}
 
 	@GetMapping("/delete/{id}")
-	public String deleteUser(@PathVariable("id") long id, Model model) {
+	public String deleteUser(@PathVariable("id") int id, Model model) {
 		User user = userRepository.findById(id)
 				.orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
 		userRepository.delete(user);
